+++ src/main/scala-2.11/markets/engines/orderbooks/immutable/ImmutableOrderBook.scala
@@ -13,7 +13,23 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 */
-package markets.orderbooks
+package markets.engines.orderbooks.immutable
+
+import markets.orders.Order
+
+import scala.collection.immutable
 
 
-package object immutable
+/** Base trait for all immutable order books.
+  *
+  * An order book is a collection of orders (typically either ask or bid orders).
+  *
+  * @tparam A the type of orders stored in the order book.
+  * @tparam B the type of underlying collection used to store the orders.
+  */
+trait ImmutableOrderBook[A <: Order, B <: immutable.Iterable[A]] {
+
+  protected var backingStore: B
+
+}
+
