+++ src/main/scala-2.11/markets/engines/orderbooks/immutable/ImmutableOrderBooks.scala
@@ -13,18 +13,19 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 */
-package markets.orderbooks.immutable
+package markets.engines.orderbooks.immutable
 
-import markets.orders.BidOrder
+import markets.engines.orderbooks.OrderBooks
+import markets.orders.{AskOrder, BidOrder}
 
-import scala.collection.immutable.TreeSet
+import scala.collection.immutable
 
 
-class ImmutableTreeSetBidOrderBook(implicit ordering: Ordering[BidOrder])
-  extends ImmutableTreeSetOrderBook[BidOrder] {
+/** Mixin trait providing immutable order books.
+  *
+  * @tparam A the type of orders stored in the order book.
+  * @tparam B the type of underlying immutable collection used to store the orders.
+  */
+trait ImmutableOrderBooks[A <: immutable.Iterable[AskOrder], B <: immutable.Iterable[BidOrder]]
+  extends OrderBooks[A, B]
 
-  protected var backingStore: TreeSet[BidOrder] = {
-    TreeSet.empty[BidOrder]
-  }
-
-}
