+++ src/main/scala-2.11/markets/engines/orderbooks/immutable/package.scala
@@ -13,19 +13,7 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 */
-package markets.orderbooks.immutable
-
-import markets.orderbooks.OrderBooks
-import markets.orders.{AskOrder, BidOrder}
-
-import scala.collection.immutable
+package markets.engines.orderbooks
 
 
-/** Mixin trait providing immutable order books.
-  *
-  * @tparam A the type of orders stored in the order book.
-  * @tparam B the type of underlying immutable collection used to store the orders.
-  */
-trait ImmutableOrderBooks[A <: immutable.Iterable[AskOrder], B <: immutable.Iterable[BidOrder]]
-  extends OrderBooks[A, B]
-
+package object immutable
