+++ src/main/scala-2.11/markets/engines/orderbooks/mutable/MutableOrderBook.scala
@@ -13,11 +13,24 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 */
-package markets.orderbooks
+package markets.engines.orderbooks.mutable
 
-/**
-  * Created by drpugh on 4/25/16.
+import markets.engines.orderbooks.OrderBook
+import markets.orders.Order
+
+import scala.collection.mutable
+
+
+/** Base trait for all order books.
+  *
+  * An order book is a collection of orders (typically either ask or bid orders).
+  *
+  * @tparam A the type of orders stored in the order book.
+  * @tparam B the type of underlying collection used to store the orders.
   */
-package object mutable {
+trait MutableOrderBook[A <: Order, B <: mutable.Iterable[A]] extends OrderBook[A, B] {
+
+  protected val backingStore: B
 
 }
+
