+++ src/main/scala-2.11/markets/engines/orderbooks/mutable/package.scala
@@ -13,23 +13,11 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 */
-package markets.orderbooks.immutable
+package markets.engines.orderbooks
 
-import markets.orders.Order
-
-import scala.collection.immutable
-
-
-/** Base trait for all immutable order books.
-  *
-  * An order book is a collection of orders (typically either ask or bid orders).
-  *
-  * @tparam A the type of orders stored in the order book.
-  * @tparam B the type of underlying collection used to store the orders.
+/**
+  * Created by drpugh on 4/25/16.
   */
-trait ImmutableOrderBook[A <: Order, B <: immutable.Iterable[A]] {
-
-  protected var backingStore: B
+package object mutable {
 
 }
-
